Split the server into modules
