+++ src/server/.gitignore
@@ -1,0 +1,1 @@
+npm-debug.log
